move _interpolate_input_and_output() into base runner class
